Move statement splitter to core module
