+++ sadu-core/src/test/java/de/chojo/sadu/updater/StatementSplitterTest.java
@@ -4,8 +4,9 @@
  *     Copyright (C) RainbowDashLabs and Contributor
  */
 
-package de.chojo.sadu.databases;
+package de.chojo.sadu.updater;
 
+import de.chojo.sadu.updater.StatementSplitter;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
